requirements.txt
discord.py==2.3.2
groq==0.4.2
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
opencv-python==4.9.0.80
Pillow==10.2.0
ImageHash==4.3.1
python-dotenv==1.0.0